Outsourced Ports and Encryption
