+++ src/handleEncryption.h
@@ -1,0 +1,4 @@
+#ifndef handleEncryption_h
+#define handleEncryption_h
+
+#endif
